requirements.txt
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cycler==0.11.0
docopt==0.6.2
flake8==4.0.1
fonttools==4.28.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
numpy==1.21.4
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
pipreqs==0.4.11
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
regex==2021.11.10
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
yarg==0.1.9
